Using uint64 for timestamps

+++ test/helpers/FinalizableCrowdsaleImpl.sol
@@ -7,8 +7,8 @@ import '../../contracts/crowdsale/FinalizableCrowdsale.sol';
 contract FinalizableCrowdsaleImpl is FinalizableCrowdsale {
 
   function FinalizableCrowdsaleImpl (
-    uint256 _startTime,
-    uint256 _endTime,
+    uint64 _startTime,
+    uint64 _endTime,
     uint256 _rate,
     address _wallet
   ) public
